Refactor MainActivity and HomeActivity to ViewModel
