+++ app/build.gradle
@@ -31,6 +31,7 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.0-beta2'
     implementation 'com.google.android.material:material:1.1.0-beta01'
+    implementation "androidx.lifecycle:lifecycle-extensions:$lifecycle_version"
 
     //Coorutine
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.0'
