+++ build.gradle
@@ -2,6 +2,7 @@
 
 buildscript {
     ext.kotlin_version = '1.3.50'
+    ext.lifecycle_version = "2.1.0"
 
     repositories {
         google()
